refactor: abstract the command for linux, and change the commands names
